HiSeCu Security related code snippet
This code can be used to monitor security patches that must be applied on SLES (SuSE Linux Enterprise) Servers with Xymon.
It may be useful for other types of SuSe Linux servers as well, or other monitoring tools.

Often, software is too tailor-made to be useful for others, but snippets of the code could be useful.

(C) 2013 Nomeka [tm] / HiSeCU, CC-By-SA 3.0

Feel free to use this code in your software, it is free, as long as you keep the comments in place, including the copyright notice

Carel Fenijn <snippets@hisecu.org>, May 2013.