HiSeCu Security related code snippet
This code can be used to monitor auditd related things on linux servers. It was written for use
on SLES systems in Xymon plugins, but may be useful for other distro's and monitoring tools as well.

Often, software is too tailor-made to be useful for others, but snippets of the code could be useful.

(C) 2013 Nomeka [tm] / HiSeCU, CC-By-SA 3.0

Feel free to use this code in your software, it is free, as long as you keep the comments in place, including the copyright notice

Carel Fenijn <snippets@hisecu.org>, May 2013.